moved all home compoennts into it's folder -home- and updated the imports
