+++ components/home/AboutPrevious.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/no-unescaped-entities */
 import Link from "next/link";
-import Card from "./ui/Card";
-import { FlexContainer, SpacerY } from "./ui/Spacer";
-import { Subtitle, Text } from "./ui/Typography";
+import Card from "@/components/ui/Card";
+import { FlexContainer, SpacerY } from "@/components/ui/Spacer";
+import { Subtitle, Text } from "@/components/ui/Typography";
 
 export default function AboutPrevious() {
   return (
